GeneFinder.Codon — Type

Codon <: BioSequence{DNAAlphabet{2}}

A Struct representing a codon, which is a subtype of BioSequence with an Alphabet of type DNAAlphabet{2}. It has a single field x of type UInt8. This was implemente in The following implementation is from https://biojulia.net/BioSequences.jl/stable/interfaces/

GeneFinder.ORF — Type

struct ORF
    location::UnitRange{Int64}
    strand::Char

    ORF(location, strand) = new(location, strand)
end

The ORF struct represents an open reading frame in a DNA sequence. It has two fields:

location: which is a UnitRange{Int64} indicating the start and end locations of the ORF in the sequence

strand: is a Char type indicating whether the ORF is on the forward ('+') or reverse ('-') strand of the sequence.

GeneFinder.cdsgenerator — Method

cdsgenerator(sequence::LongDNA)

A function to generete CDSs sequence out of a DNA sequence.

The cdsgenerator is a generator function that takes a LongDNA sequence and returns an iterator over the given sequence, containing the coding sequences (CDSs) found in the sequence and the ORF. It uses the simplefinder function to find open reading frames (ORFs) in the sequence, and then it extracts the actual CDS sequence from each ORF. The function also searches the reverse complement of the sequence, so it finds CDSs on both strands.

GeneFinder.locationgenerator — Method

locationgenerator(sequence::LongDNA)

Generate the locations of ORFs in the given DNA sequence.

This function searches the sequence for start codons, and generates ranges of indices corresponding to the locations of ORFs in the sequence. The ORFs are generated by iterating over the start codon indices and searching for the first stop codon that follows each start codon. ORFs that contain premature stop codons are filtered out using the hasprematurestop function.

The sequence argument must be a LongDNA object, which is a type of DNA sequence with a longer maximum length than the DNA type.

Returns: A generator expression that yields ranges of indices corresponding to the locations of ORFs in the sequence.

GeneFinder.orfgenerator — Method

orfgenerator(sequence::LongDNA)

Generate ORFs from the given DNA sequence.

This function generates ORFs from the forward and reverse complement strands of the sequence using the locationgenerator function. It generates an ORF object for each range of indices returned by locationgenerator, and includes a '+' or '-' strand label to indicate the strand from which the ORF was generated.

The sequence argument must be a LongDNA object, which is a type of DNA sequence with a longer maximum length than the DNA type.

Returns: A generator expression that yields ORF objects corresponding to the ORFs in the sequence.

GeneFinder.proteingenerator — Method

proteingenerator(sequence::LongDNA)

As its name suggest this generator function that iterates over the sequence to find proteins directly from a DNA sequence. The cdsgenerator function takes a LongDNA sequence and returns a Vector{CDS} containing the coding sequences (CDSs) found in the sequence.

GeneFinder.simplefinder — Method

simplefinder(sequence::LongDNA)

The simplest algorithm that finds ORFs in a DNA sequence.

The simplefinder function takes a LongDNA sequence and returns a Vector{ORF} containing the ORFs found in the sequence. It searches the sequence for start codons (ATG) and stops when it finds a stop codon (TAG, TAA, or TGA), adding each ORF it finds to the vector. The function also searches the reverse complement of the sequence, so it finds ORFs on both strands. This function has not ORFs size and overlapping condition contraints. Thus it might consider aa"M*" a posible encoding protein from the resulting ORFs.