GeneFinder.ORF

GeneFinder.fasta_to_dna

GeneFinder.findorfs

GeneFinder.get_orfs_aa

GeneFinder.get_orfs_dna

GeneFinder.hasprematurestop

GeneFinder.locationiterator

GeneFinder.record_orfs_faa

GeneFinder.record_orfs_fna

GeneFinder.write_orfs_bed

GeneFinder.write_orfs_faa

GeneFinder.write_orfs_fna

GeneFinder.write_orfs_gff

GeneFinder.ORF — Type

struct ORF
    location::UnitRange{Int64}
    strand::Char
    frame::Int
end

The ORF struct represents an open reading frame in a DNA sequence. It has two fields:

location: which is a UnitRange{Int64} indicating the start and end locations of the ORF in the sequence

strand: is a Char type indicating whether the ORF is on the forward ('+') or reverse ('-') strand of the sequence.

GeneFinder.fasta_to_dna — Method

fasta_to_dna(input::String)

Converts a FASTA formatted file (even if it is a multi-fasta) to an array of LongSequence{DNAAlphabet{4}} objects.

GeneFinder.findorfs — Method

findorfs(sequence::NucleicAlphabet{DNAAlphabet{N}}; alternative_start::Bool=false, min_len::Int64=6)::Vector{ORF} where {N}

A simple implementation that finds ORFs in a DNA sequence.

The findorfs function takes a LongSequence{DNAAlphabet{4}} sequence and returns a Vector{ORF} containing the ORFs found in the sequence. It searches entire regularly expressed CDS, adding each ORF it finds to the vector. The function also searches the reverse complement of the sequence, so it finds ORFs on both strands. Extending the starting codons with the alternative_start = true will search for ATG, GTG, and TTG. Some studies have shown that in E. coli (K-12 strain), ATG, GTG and TTG are used 83 %, 14 % and 3 % respectively.

Note

This function has not ORFs scoring scheme. Thus it might consider aa"M*" a posible encoding protein from the resulting ORFs.

Keywords

alternative_start::Bool=false: If true will pass the extended start codons to search. This will increase 3x the exec. time.

min_len::Int64=6: Length of the allowed ORF. Default value allow aa"M*" a posible encoding protein from the resulting ORFs.

GeneFinder.get_orfs_aa — Method

get_orfs_aa(sequence::NucleicSeqOrView{DNAAlphabet{N}}; kwargs...) where {N}

This function takes a NucleicSeqOrView{DNAAlphabet{N}} sequence and identifies the open reading frames (ORFs) using the findorfs() function. The function then translates each ORF into an amino acid sequence and stores it in a Vector{LongSubSeq{AminoAcidAlphabet}}.

Arguments

sequence: The input sequence as a NucleicSeqOrView{DNAAlphabet{N}}

Keyword Arguments

alternative_start::Bool=false: If set to true, the function considers alternative start codons when searching for ORFs. This increases the execution time by approximately 3x.

min_len::Int64=6: The minimum length of the allowed ORF. By default, it allows ORFs that can encode at least one amino acid (e.g., aa"M*").

GeneFinder.get_orfs_dna — Method

get_orfs_dna(sequence::NucleicSeqOrView{DNAAlphabet{N}}, output::String; kwargs...) where {N}

This function takes a NucleicSeqOrView{DNAAlphabet{N}} sequence and identifies the open reading frames (ORFs) using the findorfs() function. The function then extracts the DNA sequence of each ORF and stores it in a Vector{LongSubSeq{DNAAlphabet{4}}}.

Arguments

sequence: The input sequence as a NucleicSeqOrView{DNAAlphabet{N}}

Keyword Arguments

alternative_start::Bool=false: If set to true, the function considers alternative start codons when searching for ORFs. This increases the execution time by approximately 3x.

min_len::Int64=6: The minimum length of the allowed ORF. By default, it allows ORFs that can encode at least one amino acid (e.g., aa"M*").

GeneFinder.hasprematurestop — Method

hasprematurestop(sequence::LongNucOrView{4})::Bool

Determine whether the sequence of type LongSequence{DNAAlphabet{4}} contains a premature stop codon.

Returns a boolean indicating whether the sequence has more than one stop codon.

GeneFinder.locationiterator — Method

locationiterator(sequence::NucleicSeqOrView{DNAAlphabet{N}}; alternative_start::Bool=false) where {N}

This is an iterator function that uses regular expressions to search the entire ORF (instead of start and stop codons) in a LongSequence{DNAAlphabet{4}} sequence. It uses an anonymous function that will find the first regularly expressed ORF. Then using this anonymous function it creates an iterator that will apply it until there is no other CDS.

Note

As a note of the implementation we want to expand on how the ORFs are found:

The expression (?:[N]{3})*? serves as the boundary between the start and stop codons. Within this expression, the character class [N]{3} captures exactly three occurrences of any character (representing nucleotides using IUPAC codes). This portion functions as the regular codon matches. Since it is enclosed within a non-capturing group (?:) and followed by *?, it allows for the matching of intermediate codons, but with a preference for the smallest number of repetitions.

In summary, the regular expression ATG(?:[N]{3})*?T(AG|AA|GA) identifies patterns that start with "ATG," followed by any number of three-character codons (represented by "N" in the IUPAC code), and ends with a stop codon "TAG," "TAA," or "TGA." This pattern is commonly used to identify potential protein-coding regions within genetic sequences.

See more about the discussion here

GeneFinder.record_orfs_faa — Method

record_orfs_faa(sequence::NucleicSeqOrView{DNAAlphabet{N}}; kwargs...) where {N}

This function takes a nucleic acid sequence and finds all open reading frames (ORFs) in the sequence. An ORF is a sequence of DNA that starts with a start codon and ends with a stop codon. The function returns a list of FASTA records, where each record represents an ORF in the sequence.

Arguments

sequence: The nucleic acid sequence to search for ORFs.

alternative_start: A boolean indicating whether to consider alternative start codons. Default is false.

code: The genetic code to use for translation. Default is the NCBI translation table 1.

min_len: The minimum length of an ORF. ORFs shorter than this length will be ignored. Default is 6.

Returns

A list of FASTA records representing the ORFs found in the sequence.

GeneFinder.record_orfs_fna — Method

record_orfs_fna(sequence::NucleicSeqOrView{DNAAlphabet{N}}; kwargs...) where {N}

Record Open Reading Frames (ORFs) in a nucleic acid sequence.

Arguments

sequence::NucleicSeqOrView{DNAAlphabet{N}}: The nucleic acid sequence to search for ORFs.

alternative_start::Bool=false: If true, consider alternative start codons for ORFs.

min_len::Int=6: The minimum length of an ORF to be recorded.

Returns

An array of FASTARecord objects representing the identified ORFs.

Description

This function searches for Open Reading Frames (ORFs) in a given nucleic acid sequence. An ORF is a sequence of DNA that starts with a start codon and ends with a stop codon, without any other stop codons in between. By default, only the standard start codon (ATG) is considered, but if alternative_start is set to true, alternative start codons are also considered. The minimum length of an ORF to be recorded can be specified using the min_len argument.

GeneFinder.write_orfs_bed — Method

write_orfs_bed(input::NucleicAcidAlphabet{DNAAlphabet{N}}, output::Union{IOStream, IOBuffer}; kwargs...) where {N}
write_orfs_bed(input::NucleicAcidAlphabet{DNAAlphabet{N}}, output::String; kwargs...) where {N}
write_orfs_bed(input::String, output::Union{IOStream, IOBuffer}; kwargs...)
write_orfs_bed(input::String, output::String; kwargs...)

Write BED data to a file.

Arguments

input::NucleicAcidAlphabet{DNAAlphabet{N}}: The input DNA sequence.

output::String: The output file path.

alternative_start::Bool=false: If true, alternative start codons will be used when identifying CDSs. Default is false.

min_len::Int64=6: The minimum length that a CDS must have in order to be included in the output file. Default is 6.

GeneFinder.write_orfs_faa — Method

write_orfs_faa(input::LongSequence{DNAAlphabet{4}}, output::Union{IOStream, IOBuffer}; kwargs...) where {N}
write_orfs_faa(input::LongSequence{DNAAlphabet{4}}, output::String; kwargs...) where {N}
write_orfs_faa(input::String, output::Union{IOStream, IOBuffer}; kwargs...)
write_orfs_faa(input::String, output::String; kwargs...)

Write the protein sequences encoded by the coding sequences (CDSs) of a given DNA sequence to the specified file.

Keywords

code::GeneticCode=BioSequences.standard_genetic_code: The genetic code by which codons will be translated. See BioSequences.ncbi_trans_table for more info.

alternative_start::Bool=false: If true will pass the extended start codons to search. This will increase 3x the exec. time.

min_len::Int64=6: Length of the allowed ORF. Default value allow aa"M*" a posible encoding protein from the resulting ORFs.

Examples

filename = "output.faa"

seq = dna"ATGATGCATGCATGCATGCTAGTAACTAGCTAGCTAGCTAGTAA"

open(filename, "w") do file
     write_orfs_faa(seq, file)
end

GeneFinder.write_orfs_fna — Method

write_orfs_fna(input::NucleicSeqOrView{DNAAlphabet{N}}, output::Union{IOStream, IOBuffer}; kwargs...) where {N}
write_orfs_fna(input::NucleicSeqOrView{DNAAlphabet{N}}, output::String; kwargs...) where {N}
write_orfs_fna(input::String, output::Union{IOStream, IOBuffer}; kwargs...)
write_orfs_fna(input::String, output::String; kwargs...)

Write a file containing the coding sequences (CDSs) of a given DNA sequence to the specified file.

Keywords

alternative_start::Bool=false: If true, alternative start codons will be used when identifying CDSs. Default is false.

min_len::Int64=6: The minimum length that a CDS must have in order to be included in the output file. Default is 6.

Examples

filename = "output.fna"

seq = dna"ATGATGCATGCATGCATGCTAGTAACTAGCTAGCTAGCTAGTAA"

open(filename, "w") do file
     write_orfs_fna(seq, file)
end

GeneFinder.write_orfs_gff — Method

write_orfs_gff(input::NucleicSeqOrView{DNAAlphabet{N}}, output::Union{IOStream, IOBuffer}; kwargs...) where {N}
write_orfs_gff(input::NucleicSeqOrView{DNAAlphabet{N}}, output::String; kwargs...) where {N}
write_orfs_gff(input::String, output::Union{IOStream, IOBuffer}; kwargs...)
write_orfs_gff(input::String, output::String; kwargs...)

Write GFF data to a file.

Arguments

input: The input DNA sequence.

output: The output file to write the GFF data to.

alternative_start::Bool=false: If true, extended start codons will be considered during the search, increasing the execution time.

min_len::Int64=6: The minimum length of the allowed ORF. The default value allows for possible encoding proteins with the aa"M*" sequence.