GeneFinder.stopcodons — Constant

stopcodons

A constant array of Codon instances representing the three stop codons in the standard genetic code: TAG, TAA, and TGA.

GeneFinder.CDS — Type

struct CDS
    orf::ORF
    sequence::LongDNA
end

The CDS struct represents a coding sequence in a DNA sequence. It has three fields:

orf: is the basic composible type (location::UnitRange{Int}, strand::Char) displaying the location of the ORF and the associate strand: forward ('+') or reverse ('-')

sequence: a LongDNA sequence representing the actual sequence of the CDS

GeneFinder.Codon — Type

Codon <: BioSequence{DNAAlphabet{2}}

A Struct representing a codon, which is a subtype of BioSequence with an Alphabet of type DNAAlphabet{2}. It has a single field x of type UInt8. This was implemente in The following implementation is from https://biojulia.net/BioSequences.jl/stable/interfaces/

GeneFinder.Codon — Method

Codon(iterable)

Constructor function for Codon instances. Takes an iterable as input and creates a new Codon instance. It checks that the length of the iterable is 3, and if not, it throws an error. It then creates a new UInt variable x and assigns it a value of 0. It then iterates over the iterable and encodes each element using the BioSequences.encode function and the Alphabet of the Codon type. It shifts the encoded value by 6 minus 2 times the index of the element, and then ORs it with the current value of x. Finally, it creates a new Codon instance with the value of x modulo UInt8.

GeneFinder.ORF — Type

struct ORF
    location::UnitRange{Int64}
    strand::Char

    ORF(location, strand) = new(location, strand)
end

The ORF struct represents an open reading frame in a DNA sequence. It has two fields:

location: which is a UnitRange{Int64} indicating the start and end locations of the ORF in the sequence

strand: is a Char type indicating whether the ORF is on the forward ('+') or reverse ('-') strand of the sequence.

GeneFinder.Protein — Type

struct Protein
    orf::ORF
    sequence::LongDNA
end

Similarly to the CDS struct, the Protein struct represents a encoded protein sequence in a DNA sequence. It has three fields:

orf: is the basic composible type (location::UnitRange{Int}, strand::Char) of the sequence

sequence: a LongAA sequence representing the actual translated sequence of the CDS

Base.copy — Method

copy(seq::Codon)

A Base function that takes a Codon instance seq as input and returns a new Codon instance with the same value as seq.x.

Base.length — Method

length(::Codon)

Returns the length of a Codon instance. It is defined as a constant with a value of 3.

BioSequences.extract_encoded_element — Method

extract_encoded_element(x::Codon, i::Int)

Part of the BioSequences module. Takes a Codon instance x and an integer i as input. Returns the value of x.x shifted right by 6 minus 2 times i, ANDed with 3, and then modulo UInt.

GeneFinder.findcds — Method

findcds(sequence::LongDNA)

A function to generete CDSs sequence out of a DNA sequence.

The findcds function takes a LongDNA sequence and returns a Vector{CDS} containing the coding sequences (CDSs) found in the sequence. It uses the simplefinder function to find open reading frames (ORFs) in the sequence, and then it extracts the actual CDS sequence from each ORF. The function also searches the reverse complement of the sequence, so it finds CDSs on both strands.

GeneFinder.findproteins — Method

findproteins(sequence::LongDNA)

As its name suggest this function generate the possible proteins directly from a DNA sequence. The findcds function takes a LongDNA sequence and returns a Vector{CDS} containing the coding sequences (CDSs) found in the sequence.

GeneFinder.simplefinder — Method

simplefinder(sequence::LongDNA)

The simplest algorithm that finds ORFs in a DNA sequence.

The simplefinder function takes a LongDNA sequence and returns a Vector{ORF} containing the ORFs found in the sequence. It searches the sequence for start codons (ATG) and stops when it finds a stop codon (TAG, TAA, or TGA), adding each ORF it finds to the vector. The function also searches the reverse complement of the sequence, so it finds ORFs on both strands. This function has not ORFs size and overlapping condition contraints. Thus it might consider aa"M*" a posible encoding protein from the resulting ORFs.