GeneFinder.CDS — Type

struct CDS
    orf::ORF
    sequence::LongSubSeq{DNAAlphabet{4}}
end

The CDS struct represents a coding sequence in a DNA sequence. It has three fields:

orf: is the basic composible type (location::UnitRange{Int}, strand::Char) displaying the location of the ORF and the associate strand: forward ('+') or reverse ('-')

sequence: a LongDNA sequence representing the actual sequence of the CDS

GeneFinder.GeneFeatures — Type

struct GeneFeatures
    seqname::String
    start::Int64
    stop::Int64
    score::Float64
    strand::Char
    frame::'.'
    attribute::
end

This is the main Gene struct, based on the fields that could be found in a GFF3, still needs to be defined correctly, The idea is correct the frame and attributes that will have something like a possible list (id=Char;name=;locus_tag). The write and get functions should have a dedicated method for this struct.

GeneFinder.ORF — Type

struct ORF
    location::UnitRange{Int64}
    strand::Char
end

The ORF struct represents an open reading frame in a DNA sequence. It has two fields:

location: which is a UnitRange{Int64} indicating the start and end locations of the ORF in the sequence

strand: is a Char type indicating whether the ORF is on the forward ('+') or reverse ('-') strand of the sequence.

GeneFinder.Protein — Type

struct Protein
    sequence::LongSequence
    orf::ORF
end

Similarly to the CDS struct, the Protein struct represents a encoded protein sequence in a DNA sequence. It has three fields:

orf: is the basic composible type (location::UnitRange{Int}, strand::Char) of the sequence

sequence: a LongSequence sequence representing the actual translated sequence of the CDS

GeneFinder.cdsgenerator — Method

cdsgenerator(sequence::LongDNA; kwargs...)
cdsgenerator(sequence::String; kwargs...)

A function to generete CDSs sequence out of a DNA sequence.

The cdsgenerator is a generator function that takes a LongDNA sequence and returns an iterator over the given sequence, containing the coding sequences (CDSs) found in the sequence and the ORF. It uses the findorfs function to find open reading frames (ORFs) in the sequence, and then it extracts the actual CDS sequence from each ORF, returining both. The function also searches the reverse complement of the sequence, so it finds CDSs on both strands.

Keywords

alternative_start::Bool=false: If true will pass the extended start codons to search. This will increase 3x the exec. time.

min_len::Int64=6: Length of the allowed ORF. Default value allow aa"M*" a posible encoding protein from the resulting ORFs.

GeneFinder.eachcodon — Method

eachcodon(sequence::LongDNA)

Iterate through the codons in the sequence of type LongDNA.

Returns an iterator yielding Codon objects for each codon in the sequence.

GeneFinder.fasta_to_dna — Method

fasta_to_dna(input::String)

Converts a FASTA formatted file (even if it is a multi-fasta) to an array of LongDNA objects.

GeneFinder.findorfs — Method

findorfs(sequence::LongDNA; kwargs...)::Vector{ORF}
findorfs(sequence::String; kwargs...)::Vector{ORF}

A simple implementation that finds ORFs in a DNA sequence.

The findorfs function takes a LongDNA sequence and returns a Vector{ORF} containing the ORFs found in the sequence. It searches entire regularly expressed CDS, adding each ORF it finds to the vector. The function also searches the reverse complement of the sequence, so it finds ORFs on both strands. Extending the starting codons with the alternative_start = true will search for ATG, GTG, and TTG. Some studies have shown that in E. coli (K-12 strain), ATG, GTG and TTG are used 83 %, 14 % and 3 % respectively.

Note

This function has not ORFs scoring scheme. Thus it might consider aa"M*" a posible encoding protein from the resulting ORFs.

Keywords

alternative_start::Bool=false: If true will pass the extended start codons to search. This will increase 3x the exec. time.

min_len::Int64=6: Length of the allowed ORF. Default value allow aa"M*" a posible encoding protein from the resulting ORFs.

GeneFinder.getcds — Method

getcds(input::LongDNA, output::String; kwargs...)
getcds(input::String, output::String; kwargs...)

This function will take a LongDNA or String sequence and by means of the findorfs() function will push LongSubSeq{DNAAlphabet{4}} into a Vector{}

Keywords

alternative_start::Bool=false: If true will pass the extended start codons to search. This will increase 3x the exec. time.

min_len::Int64=6: Length of the allowed ORF. Default value allow aa"M*" a posible encoding protein from the resulting ORFs.

GeneFinder.getproteins — Method

getproteins(input::LongDNA, output::String; kwargs...)
getproteins(input::String, output::String; kwargs...)

Similar to getcds() function, it will take a LongDNA or String sequence and by means of the findorfs() and the translate() function will push LongAAs into a Vector

Keywords

alternative_start::Bool=false: If true will pass the extended start codons to search. This will increase 3x the exec. time.

min_len::Int64=6: Length of the allowed ORF. Default value allow aa"M*" a posible encoding protein from the resulting ORFs.

GeneFinder.hasprematurestop — Method

hasprematurestop(sequence::LongDNA)::Bool

Determine whether the sequence of type LongDNA contains a premature stop codon.

Returns a boolean indicating whether the sequence has more than one stop codon.

GeneFinder.locationiterator — Method

locationiterator(sequence::LongDNA; alternative_start::Bool=false)

This is an iterator function that uses regular expressions to search the entire CDS (instead of start and stop codons) in a LongDNA sequence. It uses an anonymous function that will find the first regularly expressed CDS. Then using this anonymous function it creates an iterator that will apply it until there is no other CDS.

GeneFinder.proteingenerator — Method

proteingenerator(sequence::LongDNA; kwargs...)
proteingenerator(sequence::String; kwargs...)

As its name suggest this generator function iterates over the sequence to find proteins directly from a DNA sequence. The proteingenerator function takes a LongDNA sequence and returns a Vector{CDS} containing the coding sequences (CDSs) found in the sequence and the associated ORF.

Keywords

code::GeneticCode=BioSequences.standard_genetic_code: The genetic code by which codons will be translated. See BioSequences.ncbi_trans_table for more info.

alternative_start::Bool=false: If true will pass the extended start codons to search. This will increase 3x the exec. time.

min_len::Int64=6: Length of the allowed ORF. Default value allow aa"M*" a posible encoding protein from the resulting ORFs.

GeneFinder.write_bed — Method

write_bed(input::LongDNA, output::String; kwargs...)
write_bed(input::String, output::String; kwargs...)

Write BED data to a file.

Keywords

alternative_start::Bool=false: If true will pass the extended start codons to search. This will increase 3x the exec. time.

min_len::Int64=6: Length of the allowed ORF. Default value allow aa"M*" a posible encoding protein from the resulting ORFs.

GeneFinder.write_cds — Method

write_cds(input::LongDNA, output::String; kwargs...)
write_cds(input::String, output::String; kwargs...)

Write a file containing the coding sequences (CDSs) of a given DNA sequence to the specified file.

Keywords

alternative_start: A boolean value indicating whether alternative start codons should be used when identifying CDSs. Default is false.

min_len: An integer representing the minimum length that a CDS must have in order to be included in the output file. Default is 6.

GeneFinder.write_gff — Method

write_gff(input::LongDNA, output::String; kwargs...)
write_gff(input::String, output::String; kwargs...)

Write GFF data to a file.

Keywords

alternative_start::Bool=false: If true will pass the extended start codons to search. This will increase 3x the exec. time.

min_len::Int64=6: Length of the allowed ORF. Default value allow aa"M*" a posible encoding protein from the resulting ORFs.

GeneFinder.write_proteins — Method

write_proteins(input::LongDNA, output::String; kwargs...)

Write the protein sequences encoded by the coding sequences (CDSs) of a given DNA sequence to the specified file.

Keywords

code::GeneticCode=BioSequences.standard_genetic_code: The genetic code by which codons will be translated. See BioSequences.ncbi_trans_table for more info.

alternative_start::Bool=false: If true will pass the extended start codons to search. This will increase 3x the exec. time.

min_len::Int64=6: Length of the allowed ORF. Default value allow aa"M*" a posible encoding protein from the resulting ORFs.