GeneFinder.CDS — Type

struct CDS
    location::UnitRange{Int64}
    strand::Char
    sequence::LongDNA
end

The CDS struct represents a coding sequence in a DNA sequence. It has three fields:

- `orf`: is the basic composible type (`location::UnitRange{Int}`, strand::Char) displaying the location of the ORF and the associate strand: forward ('+') or reverse ('-')
- `sequence`: a `LongDNA` sequence representing the actual sequence of the CDS

GeneFinder.ORF — Type

struct ORF
    location::UnitRange{Int64}
    strand::Char
end

The ORF struct represents an open reading frame in a DNA sequence. It has two fields:

- `location`: which is a UnitRange{Int64} indicating the start and end locations of the ORF in the sequence
- `strand`:  is a Char type indicating whether the ORF is on the forward ('+') or reverse ('-') strand of the sequence.

GeneFinder.Protein — Type

struct Protein
    location::UnitRange{Int64}
    strand::Char
    sequence::LongDNA
end

Similarly to the CDS struct, the Protein struct represents a encoded protein sequence in a DNA sequence. It has three fields:

- `orf`: is the basic composible type (`location::UnitRange{Int}`, strand::Char) of the sequence
- `sequence`: a `LongAA` sequence representing the actual translated sequence of the CDS

GeneFinder.findcds — Method

`findcds(sequence::LongDNA)`

A function to generete CDSs sequence out of a DNA sequence.

The findcds function takes a LongDNA sequence and returns a Vector{CDS} containing the coding sequences (CDSs) found in the sequence. It uses the simplefinder function to find open reading frames (ORFs) in the sequence, and then it extracts the actual CDS sequence from each ORF. The function also searches the reverse complement of the sequence, so it finds CDSs on both strands.

GeneFinder.findproteins — Method

`findproteins(sequence::LongDNA)`

As its name suggest this function generate the possible proteins directly from a DNA sequence. The findcds function takes a LongDNA sequence and returns a Vector{CDS} containing the coding sequences (CDSs) found in the sequence.

GeneFinder.simplefinder — Method

`simplefinder(sequence::LongDNA)`

The simplest algorithm that finds ORFs in a DNA sequence.

The simplefinder function takes a LongDNA sequence and returns a Vector{ORF} containing the ORFs found in the sequence. It searches the sequence for start codons (ATG) and stops when it finds a stop codon (TAG, TAA, or TGA), adding each ORF it finds to the vector. The function also searches the reverse complement of the sequence, so it finds ORFs on both strands. This function has not ORFs size and overlapping condition contraints. Thus it might consider aa"M*" a posible encoding protein from the resulting ORFs.