GeneFinder.GeneFeatures

GeneFinder.ORF

GeneFinder.fasta_to_dna

GeneFinder.findorfs

GeneFinder.getorfaa

GeneFinder.getorfdna

GeneFinder.hasprematurestop

GeneFinder.locationiterator

GeneFinder.nucleotidefreqs

GeneFinder.write_bed

GeneFinder.write_cds

GeneFinder.write_gff

GeneFinder.write_proteins

GeneFinder.GeneFeatures — Type

struct GeneFeatures
    seqname::String
    start::Int64
    stop::Int64
    score::Float64
    strand::Char
    frame::'.' 
    attribute::
end

This is the main Gene struct, based on the fields that could be found in a GFF3, still needs to be defined correctly, The idea is correct the frame and attributes that will have something like a possible list (id=Char;name=;locus_tag). The write and get functions should have a dedicated method for this struct.

GeneFinder.ORF — Type

struct ORF
    location::UnitRange{Int64}
    strand::Char
    frame::Integer
end

The ORF struct represents an open reading frame in a DNA sequence. It has two fields:

location: which is a UnitRange{Int64} indicating the start and end locations of the ORF in the sequence

strand: is a Char type indicating whether the ORF is on the forward ('+') or reverse ('-') strand of the sequence.

GeneFinder.fasta_to_dna — Method

fasta_to_dna(input::String)

Converts a FASTA formatted file (even if it is a multi-fasta) to an array of LongSequence{DNAAlphabet{4}} objects.

GeneFinder.findorfs — Method

findorfs(sequence::LongSequence{DNAAlphabet{4}}; kwargs...)::Vector{ORF}
findorfs(sequence::String; kwargs...)::Vector{ORF}

A simple implementation that finds ORFs in a DNA sequence.

The findorfs function takes a LongSequence{DNAAlphabet{4}} sequence and returns a Vector{ORF} containing the ORFs found in the sequence. It searches entire regularly expressed CDS, adding each ORF it finds to the vector. The function also searches the reverse complement of the sequence, so it finds ORFs on both strands. Extending the starting codons with the alternative_start = true will search for ATG, GTG, and TTG. Some studies have shown that in E. coli (K-12 strain), ATG, GTG and TTG are used 83 %, 14 % and 3 % respectively.

Note

This function has not ORFs scoring scheme. Thus it might consider aa"M*" a posible encoding protein from the resulting ORFs.

Keywords

alternative_start::Bool=false: If true will pass the extended start codons to search. This will increase 3x the exec. time.

min_len::Int64=6: Length of the allowed ORF. Default value allow aa"M*" a posible encoding protein from the resulting ORFs.

GeneFinder.getorfaa — Method

getorfaa(input::LongSequence{DNAAlphabet{4}}; kwargs...)
getorfaa(input::String; kwargs...) ## for strings per se

This function takes a LongSequence{DNAAlphabet{4}} or String sequence and identifies the open reading frames (ORFs) using the findorfs() function. The function then translates each ORF into an amino acid sequence and stores it in a Vector{LongSubSeq{AminoAcidAlphabet}}.

Arguments

input: The input sequence as a LongSequence{DNAAlphabet{4}} or String.

Keyword Arguments

alternative_start::Bool=false: If set to true, the function considers alternative start codons when searching for ORFs. This increases the execution time by approximately 3x.

min_len::Int64=6: The minimum length of the allowed ORF. By default, it allows ORFs that can encode at least one amino acid (e.g., aa"M*").

GeneFinder.getorfdna — Method

getorfdna(input::LongSequence{DNAAlphabet{4}}, output::String; kwargs...)
getorfdna(input::String; kwargs...) ## for strings per se

This function takes a LongSequence{DNAAlphabet{4}} or String sequence and identifies the open reading frames (ORFs) using the findorfs() function. The function then extracts the DNA sequence of each ORF and stores it in a Vector{LongSubSeq{DNAAlphabet{4}}}.

Arguments

input: The input sequence as a LongSequence{DNAAlphabet{4}} or String.

Keyword Arguments

alternative_start::Bool=false: If set to true, the function considers alternative start codons when searching for ORFs. This increases the execution time by approximately 3x.

min_len::Int64=6: The minimum length of the allowed ORF. By default, it allows ORFs that can encode at least one amino acid (e.g., aa"M*").

GeneFinder.hasprematurestop — Method

hasprematurestop(sequence::LongNucOrView{4})::Bool

Determine whether the sequence of type LongSequence{DNAAlphabet{4}} contains a premature stop codon.

Returns a boolean indicating whether the sequence has more than one stop codon.

GeneFinder.locationiterator — Method

locationiterator(sequence::LongSequence{DNAAlphabet{4}}; alternative_start::Bool=false)

This is an iterator function that uses regular expressions to search the entire ORF (instead of start and stop codons) in a LongSequence{DNAAlphabet{4}} sequence. It uses an anonymous function that will find the first regularly expressed ORF. Then using this anonymous function it creates an iterator that will apply it until there is no other CDS.

Note

As a note of the implementation we want to expand on how the ORFs are found:

The expression (?:[N]{3})*? serves as the boundary between the start and stop codons. Within this expression, the character class [N]{3} captures exactly three occurrences of any character (representing nucleotides using IUPAC codes). This portion functions as the regular codon matches. Since it is enclosed within a non-capturing group (?:) and followed by *?, it allows for the matching of intermediate codons, but with a preference for the smallest number of repetitions.

In summary, the regular expression ATG(?:[N]{3})*?T(AG|AA|GA) identifies patterns that start with "ATG," followed by any number of three-character codons (represented by "N" in the IUPAC code), and ends with a stop codon "TAG," "TAA," or "TGA." This pattern is commonly used to identify potential protein-coding regions within genetic sequences.

See more about the discussion here

GeneFinder.nucleotidefreqs — Method

nucleotidefreqs(sequence::LongSequence{DNAAlphabet{4}}) -> Dict{DNA, Float64}

Calculate the frequency of each nucleotide in a DNA sequence.

Arguments

sequence::LongSequence{DNAAlphabet{4}}: A LongSequence{DNAAlphabet{4}} sequence.

Returns

A dictionary with each nucleotide in the sequence as a key, and its frequency as a value.

Example

seq = dna"CCTCCCGGACCCTGGGCTCGGGAC"

nucleotidefreqs(seq)

Dict{DNA, Float64} with 4 entries:
DNA_T => 0.125
DNA_A => 0.0833333
DNA_G => 0.333333
DNA_C => 0.458333

GeneFinder.write_bed — Method

write_bed(input::LongSequence{DNAAlphabet{4}}, output::String; kwargs...)
write_bed(input::String, output::String; kwargs...)

Write BED data to a file.

Keywords

alternative_start::Bool=false: If true will pass the extended start codons to search. This will increase 3x the exec. time.

min_len::Int64=6: Length of the allowed ORF. Default value allow aa"M*" a posible encoding protein from the resulting ORFs.

GeneFinder.write_cds — Method

write_cds(input::LongSequence{DNAAlphabet{4}}, output::String; kwargs...)
write_cds(input::String, output::String; kwargs...)

Write a file containing the coding sequences (CDSs) of a given DNA sequence to the specified file.

Keywords

alternative_start: A boolean value indicating whether alternative start codons should be used when identifying CDSs. Default is false.

min_len: An integer representing the minimum length that a CDS must have in order to be included in the output file. Default is 6.

GeneFinder.write_gff — Method

write_gff(input::LongSequence{DNAAlphabet{4}}, output::String; kwargs...)
write_gff(input::String, output::String; kwargs...)

Write GFF data to a file.

Keywords

alternative_start::Bool=false: If true will pass the extended start codons to search. This will increase 3x the exec. time.

min_len::Int64=6: Length of the allowed ORF. Default value allow aa"M*" a posible encoding protein from the resulting ORFs.

GeneFinder.write_proteins — Method

write_proteins(input::LongSequence{DNAAlphabet{4}}, output::String; kwargs...)

Write the protein sequences encoded by the coding sequences (CDSs) of a given DNA sequence to the specified file.

Keywords

code::GeneticCode=BioSequences.standard_genetic_code: The genetic code by which codons will be translated. See BioSequences.ncbi_trans_table for more info.

alternative_start::Bool=false: If true will pass the extended start codons to search. This will increase 3x the exec. time.

min_len::Int64=6: Length of the allowed ORF. Default value allow aa"M*" a posible encoding protein from the resulting ORFs.