GeneFinder.CDS — Type

struct CDS
    orf::ORF
    sequence::LongSubSeq{DNAAlphabet{4}}
end

The CDS struct represents a coding sequence in a DNA sequence. It has three fields:

orf: is the basic composible type (location::UnitRange{Int}, strand::Char) displaying the location of the ORF and the associate strand: forward ('+') or reverse ('-')

sequence: a LongSequence{DNAAlphabet{4}} sequence representing the actual sequence of the CDS

GeneFinder.GeneFeatures — Type

struct GeneFeatures
    seqname::String
    start::Int64
    stop::Int64
    score::Float64
    strand::Char
    frame::'.'
    attribute::
end

This is the main Gene struct, based on the fields that could be found in a GFF3, still needs to be defined correctly, The idea is correct the frame and attributes that will have something like a possible list (id=Char;name=;locus_tag). The write and get functions should have a dedicated method for this struct.

GeneFinder.ORF — Type

struct ORF
    location::UnitRange{Int64}
    strand::Char
end

The ORF struct represents an open reading frame in a DNA sequence. It has two fields:

location: which is a UnitRange{Int64} indicating the start and end locations of the ORF in the sequence

strand: is a Char type indicating whether the ORF is on the forward ('+') or reverse ('-') strand of the sequence.

GeneFinder.Protein — Type

struct Protein
    sequence::LongSequence
    orf::ORF
end

Similarly to the CDS struct, the Protein struct represents a encoded protein sequence in a DNA sequence. It has three fields:

orf: is the basic composible type (location::UnitRange{Int}, strand::Char) of the sequence

sequence: a LongSequence sequence representing the actual translated sequence of the CDS

GeneFinder.TCM — Type

TCM(alphabet::Vector{DNA})

A data structure for storing a DNA Transition Count Matrix (TCM). The TCM is a square matrix where each row and column corresponds to a nucleotide in the given alphabet. The value at position (i, j) in the matrix represents the number of times that nucleotide i is immediately followed by nucleotide j in a DNA sequence.

Fields:

order::Dict{DNA, Int64}: A dictionary that maps each nucleotide in the alphabet to its corresponding index in the matrix.

counts::Matrix{Int64}: The actual matrix of counts.

Internal function:

TCM(alphabet::Vector{DNA}): Constructs a new TCM object with the given alphabet. This function initializes the order field by creating a dictionary that maps each nucleotide in the alphabet to its corresponding index in the matrix. It also initializes the counts field to a matrix of zeros with dimensions len x len, where len is the length of the alphabet.

Example usage:

alphabet = [DNA_A, DNA_C, DNA_G, DNA_T]
dtcm = TCM(alphabet)

GeneFinder.TPM — Type

TPM(alphabet::Vector{DNA})

A data structure for storing a DNA Transition Probability Matrix (TPM). The TPM is a square matrix where each row and column corresponds to a nucleotide in the given alphabet. The value at position (i, j) in the matrix represents the probability of transitioning from nucleotide i to nucleotide j in a DNA sequence.

Fields:

order::Dict{DNA, Int64}: A dictionary that maps each nucleotide in the alphabet to its corresponding index in the matrix.

probabilities::Matrix{Float64}: The actual matrix of probabilities.

Example usage:

alphabet = [DNA_A, DNA_C, DNA_G, DNA_T]
dtpm = TPM(alphabet)

GeneFinder.TransitionModel — Type

struct TransitionModel

The TransitionModel struct represents a transition model with coding and noncoding matrices, and coding and noncoding initial distributions.

Fields

coding::Matrix{Float64}: The coding matrix representing transition probabilities.

noncoding::Matrix{Float64}: The noncoding matrix representing transition probabilities.

codinginits::Matrix{Float64}: The coding initial distribution.

noncodinginits::Matrix{Float64}: The noncoding initial distribution.

Constructors

TransitionModel(coding::Matrix{Float64}, noncoding::Matrix{Float64}, codinginits::Matrix{Float64}, noncodinginits::Matrix{Float64}): Create a TransitionModel instance with the specified coding and noncoding matrices and initial distributions.

TransitionModel(codingseq, noncodingseq): Create a TransitionModel instance by computing the transition probabilities and initial distributions from the given coding and noncoding sequences.

GeneFinder.cdsgenerator — Method

cdsgenerator(sequence::LongSequence{DNAAlphabet{4}}; kwargs...)
cdsgenerator(sequence::String; kwargs...)

A function to generete CDSs sequence out of a DNA sequence.

The cdsgenerator is a generator function that takes a LongSequence{DNAAlphabet{4}} sequence and returns an iterator over the given sequence, containing the coding sequences (CDSs) found in the sequence and the ORF. It uses the findorfs function to find open reading frames (ORFs) in the sequence, and then it extracts the actual CDS sequence from each ORF, returining both. The function also searches the reverse complement of the sequence, so it finds CDSs on both strands.

Keywords

alternative_start::Bool=false: If true will pass the extended start codons to search. This will increase 3x the exec. time.

min_len::Int64=6: Length of the allowed ORF. Default value allow aa"M*" a posible encoding protein from the resulting ORFs.

GeneFinder.dinucleotides — Method

dinucleotides(sequence::LongSequence{DNAAlphabet{4}})

Compute the transition counts of each dinucleotide in a given DNA sequence.

Arguments

sequence::LongSequence{DNAAlphabet{4}}: a LongSequence{DNAAlphabet{4}} object representing the DNA sequence.

Keywords

extended_alphabet::Bool=false: If true will pass the extended alphabet of DNA to search

Returns

A dictionary with keys being LongSequence{DNAAlphabet{4}} objects representing the dinucleotides, and values being the number of occurrences of each dinucleotide in the sequence.

Example

    seq = dna"AGCTAGCTAGCT"

    dinucleotides(seq)

    Dict{LongSequence{DNAAlphabet{4}}, Int64} with 16 entries:
      GG => 0
      TC => 0
      GC => 3
      CG => 0
      CC => 0
      AG => 3
      TT => 0
      AC => 0
      TA => 2
      GT => 0
      GA => 0
      CT => 3
      CA => 0
      AT => 0
      AA => 0
      TG => 0

GeneFinder.fasta_to_dna — Method

fasta_to_dna(input::String)

Converts a FASTA formatted file (even if it is a multi-fasta) to an array of LongSequence{DNAAlphabet{4}} objects.

GeneFinder.findorfs — Method

findorfs(sequence::LongSequence{DNAAlphabet{4}}; kwargs...)::Vector{ORF}
findorfs(sequence::String; kwargs...)::Vector{ORF}

A simple implementation that finds ORFs in a DNA sequence.

The findorfs function takes a LongSequence{DNAAlphabet{4}} sequence and returns a Vector{ORF} containing the ORFs found in the sequence. It searches entire regularly expressed CDS, adding each ORF it finds to the vector. The function also searches the reverse complement of the sequence, so it finds ORFs on both strands. Extending the starting codons with the alternative_start = true will search for ATG, GTG, and TTG. Some studies have shown that in E. coli (K-12 strain), ATG, GTG and TTG are used 83 %, 14 % and 3 % respectively.

Note

This function has not ORFs scoring scheme. Thus it might consider aa"M*" a posible encoding protein from the resulting ORFs.

Keywords

alternative_start::Bool=false: If true will pass the extended start codons to search. This will increase 3x the exec. time.

min_len::Int64=6: Length of the allowed ORF. Default value allow aa"M*" a posible encoding protein from the resulting ORFs.

GeneFinder.getcds — Method

getcds(input::LongSequence{DNAAlphabet{4}}, output::String; kwargs...)
getcds(input::String, output::String; kwargs...) ## for strings per se

This function will take a LongSequence{DNAAlphabet{4}} or String sequence and by means of the findorfs() function will push LongSubSeq{DNAAlphabet{4}} into a Vector{}

Keywords

alternative_start::Bool=false: If true will pass the extended start codons to search. This will increase 3x the exec. time.

min_len::Int64=6: Length of the allowed ORF. Default value allow aa"M*" a posible encoding protein from the resulting ORFs.

GeneFinder.getproteins — Method

getproteins(input::LongSequence{DNAAlphabet{4}}, output::String; kwargs...)
getproteins(input::String, output::String; kwargs...)

Similar to getcds() function, it will take a LongSequence{DNAAlphabet{4}} or String sequence and by means of the findorfs() and the translate() function will push LongAAs into a Vector

Keywords

alternative_start::Bool=false: If true will pass the extended start codons to search. This will increase 3x the exec. time.

min_len::Int64=6: Length of the allowed ORF. Default value allow aa"M*" a posible encoding protein from the resulting ORFs.

GeneFinder.locationiterator — Method

locationiterator(sequence::LongSequence{DNAAlphabet{4}}; alternative_start::Bool=false)

This is an iterator function that uses regular expressions to search the entire CDS (instead of start and stop codons) in a LongSequence{DNAAlphabet{4}} sequence. It uses an anonymous function that will find the first regularly expressed CDS. Then using this anonymous function it creates an iterator that will apply it until there is no other CDS.

GeneFinder.nucleotidefreqs — Method

nucleotidefreqs(sequence::LongSequence{DNAAlphabet{4}}) -> Dict{DNA, Float64}

Calculate the frequency of each nucleotide in a DNA sequence.

Arguments

sequence::LongSequence{DNAAlphabet{4}}: A LongSequence{DNAAlphabet{4}} sequence.

Returns

A dictionary with each nucleotide in the sequence as a key, and its frequency as a value.

Example

    
    seq = dna"CCTCCCGGACCCTGGGCTCGGGAC"

    nucleotidefreqs(seq)

    Dict{DNA, Float64} with 4 entries:
    DNA_T => 0.125
    DNA_A => 0.0833333
    DNA_G => 0.333333
    DNA_C => 0.458333

GeneFinder.proteingenerator — Method

proteingenerator(sequence::LongSequence{DNAAlphabet{4}}; kwargs...)
proteingenerator(sequence::String; kwargs...)

As its name suggest this generator function iterates over the sequence to find proteins directly from a DNA sequence. The proteingenerator function takes a LongSequence{DNAAlphabet{4}} sequence and returns a Vector{CDS} containing the coding sequences (CDSs) found in the sequence and the associated ORF.

Keywords

code::GeneticCode=BioSequences.standard_genetic_code: The genetic code by which codons will be translated. See BioSequences.ncbi_trans_table for more info.

alternative_start::Bool=false: If true will pass the extended start codons to search. This will increase 3x the exec. time.

min_len::Int64=6: Length of the allowed ORF. Default value allow aa"M*" a posible encoding protein from the resulting ORFs.

GeneFinder.sequenceprobability — Method

sequenceprobability(sequence::LongNucOrView{4}, tpm::Matrix{Float64}, initials=Vector{Float64})

Compute the probability of a given sequence using a transition probability matrix and the initial probabilities distributions.

\[P(X_1 = i_1, \ldots, X_T = i_T) = \pi_{i_1}^{T-1} \prod_{t=1}^{T-1} a_{i_t, i_{t+1}}\]

Arguments

sequence::LongNucOrView{4}: The input sequence of nucleotides.

tpm::Matrix{Float64}: The transition probability matrix.

initials=Vector{Float64}: Optional initial state probabilities. Default is an empty vector.

Returns

probability::Float64: The probability of the input sequence.

Example

```

tpm = transitionprobabilitymatrix(dna"CCTCCCGGACCCTGGGCTCGGGAC")

4×4 Matrix{Float64}:
0.0   1.0    0.0    0.0
0.0   0.5    0.2    0.3
0.25  0.125  0.625  0.0
0.0   0.667  0.333  0.0

initials = initial_distribution(dna"CCTCCCGGACCCTGGGCTCGGGAC")

1×4 Matrix{Float64}:
0.0869565  0.434783  0.347826  0.130435

sequence = dna"CCTG"

4nt DNA Sequence:
CCTG

sequenceprobability(sequence, tpm, initials)

0.0217

GeneFinder.transition_count_matrix — Method

transition_count_matrix(sequence::LongSequence{DNAAlphabet{4}})

Compute the transition count matrix (TCM) of a given DNA sequence.

Arguments

sequence::LongSequence{DNAAlphabet{4}}: a LongSequence{DNAAlphabet{4}} object representing the DNA sequence.

Keywords

extended_alphabet::Bool=false: If true will pass the extended alphabet of DNA to search

Returns

A TCM object representing the transition count matrix of the sequence.

Example

``` seq = dna"AGCTAGCTAGCT"

tcm = transition_count_matrix(seq)

tcm.counts

4×4 Matrix{Int64}:
 0  0  3  0
 0  0  0  3
 0  3  0  0
 2  0  0  0

```

GeneFinder.transition_probability_matrix — Function

transition_probability_matrix(sequence::LongSequence{DNAAlphabet{4}})

Compute the transition probability matrix (TPM) of a given DNA sequence. Formally it construct $\hat{A}$ where:

\[a_{ij} = P(X_t = j \mid X_{t-1} = i) = \frac{{P(X_{t-1} = i, X_t = j)}}{{P(X_{t-1} = i)}}\]

Arguments

sequence::LongSequence{DNAAlphabet{4}}: a LongSequence{DNAAlphabet{4}} object representing the DNA sequence.

n::Int64=1: The order of the Markov model. That is the $\hat{A}^{n}$

Keywords

extended_alphabet::Bool=false: If true will pass the extended alphabet of DNA to search

Returns

A TPM object representing the transition probability matrix of the sequence.

Example

    seq = dna"AGCTAGCTAGCT"

    tpm = transition_probability_matrix(seq)

    tpm.probs

    4×4 Matrix{Float64}:
    0.0  0.0  1.0  0.0
    0.0  0.0  0.0  1.0
    0.0  1.0  0.0  0.0
    1.0  0.0  0.0  0.0

GeneFinder.write_bed — Method

write_bed(input::LongSequence{DNAAlphabet{4}}, output::String; kwargs...)
write_bed(input::String, output::String; kwargs...)

Write BED data to a file.

Keywords

alternative_start::Bool=false: If true will pass the extended start codons to search. This will increase 3x the exec. time.

min_len::Int64=6: Length of the allowed ORF. Default value allow aa"M*" a posible encoding protein from the resulting ORFs.

GeneFinder.write_cds — Method

write_cds(input::LongSequence{DNAAlphabet{4}}, output::String; kwargs...)
write_cds(input::String, output::String; kwargs...)

Write a file containing the coding sequences (CDSs) of a given DNA sequence to the specified file.

Keywords

alternative_start: A boolean value indicating whether alternative start codons should be used when identifying CDSs. Default is false.

min_len: An integer representing the minimum length that a CDS must have in order to be included in the output file. Default is 6.

GeneFinder.write_gff — Method

write_gff(input::LongSequence{DNAAlphabet{4}}, output::String; kwargs...)
write_gff(input::String, output::String; kwargs...)

Write GFF data to a file.

Keywords

alternative_start::Bool=false: If true will pass the extended start codons to search. This will increase 3x the exec. time.

min_len::Int64=6: Length of the allowed ORF. Default value allow aa"M*" a posible encoding protein from the resulting ORFs.

GeneFinder.write_proteins — Method

write_proteins(input::LongSequence{DNAAlphabet{4}}, output::String; kwargs...)

Write the protein sequences encoded by the coding sequences (CDSs) of a given DNA sequence to the specified file.

Keywords

code::GeneticCode=BioSequences.standard_genetic_code: The genetic code by which codons will be translated. See BioSequences.ncbi_trans_table for more info.

alternative_start::Bool=false: If true will pass the extended start codons to search. This will increase 3x the exec. time.

min_len::Int64=6: Length of the allowed ORF. Default value allow aa"M*" a posible encoding protein from the resulting ORFs.