GeneFinder.CDS — Type

struct CDS
    orf::ORF
    sequence::LongDNA
end

The CDS struct represents a coding sequence in a DNA sequence. It has three fields:

orf: is the basic composible type (location::UnitRange{Int}, strand::Char) displaying the location of the ORF and the associate strand: forward ('+') or reverse ('-')

sequence: a LongDNA sequence representing the actual sequence of the CDS

GeneFinder.Codon — Type

Codon <: BioSequence{DNAAlphabet{2}}

A Struct representing a codon, which is a subtype of BioSequence with an Alphabet of type DNAAlphabet{2}. It has a single field x of type UInt8. This was implemente in The following implementation is from https://biojulia.net/BioSequences.jl/stable/interfaces/

GeneFinder.ORF — Type

struct ORF
    location::UnitRange{Int64}
    strand::Char
    
end

The ORF struct represents an open reading frame in a DNA sequence. It has two fields:

location: which is a UnitRange{Int64} indicating the start and end locations of the ORF in the sequence

strand: is a Char type indicating whether the ORF is on the forward ('+') or reverse ('-') strand of the sequence.

GeneFinder.Protein — Type

struct Protein
    sequence::LongSequence
    orf::ORF
end

Similarly to the CDS struct, the Protein struct represents a encoded protein sequence in a DNA sequence. It has three fields:

orf: is the basic composible type (location::UnitRange{Int}, strand::Char) of the sequence

sequence: a LongSequence sequence representing the actual translated sequence of the CDS

GeneFinder.cds_generator — Method

cds_generator(sequence::LongDNA; kwargs...)
cds_generator(sequence::String; kwargs...)

A function to generete CDSs sequence out of a DNA sequence.

The cds_generator is a generator function that takes a LongDNA sequence and returns an iterator over the given sequence, containing the coding sequences (CDSs) found in the sequence and the ORF. It uses the orf_finder function to find open reading frames (ORFs) in the sequence, and then it extracts the actual CDS sequence from each ORF, returining both. The function also searches the reverse complement of the sequence, so it finds CDSs on both strands.

Keywords

alternative_start::Bool=false: If true will pass the extended start codons to search. This will increase 3x the exec. time.

min_len::Int64=6: Length of the allowed ORF. Default value allow aa"M*" a posible encoding protein from the resulting ORFs.

GeneFinder.eachcodon — Method

eachcodon(sequence::LongDNA)

Iterate through the codons in the sequence of type LongDNA.

Returns an iterator yielding Codon objects for each codon in the sequence.

GeneFinder.hasprematurestop — Method

hasprematurestop(sequence::LongDNA)::Bool

Determine whether the sequence of type LongDNA contains a premature stop codon.

Returns a boolean indicating whether the sequence has more than one stop codon.

GeneFinder.locationgenerator — Method

locationgenerator(sequence::LongDNA; alternative_start::Bool=false)

Generate the locations of ORFs in the given DNA sequence.

This function searches the sequence for start codons, and generates ranges of indices corresponding to the locations of ORFs in the sequence. The ORFs are generated by iterating over the start codon indices and searching for the first stop codon that follows each start codon. ORFs that contain premature stop codons are filtered out using the hasprematurestop function. The sequence argument must be a LongDNA object, which is a type of DNA sequence with a longer maximum length than the DNA type.

Returns: A generator expression that yields ranges of indices corresponding to the locations of ORFs in the sequence.

GeneFinder.orf_finder — Method

orf_finder(sequence::LongDNA; kwargs...)::Vector{ORF}
orf_finder(sequence::String; kwargs...)::Vector{ORF}

The simplest algorithm that finds ORFs in a DNA sequence.

The orf_finder function takes a LongDNA sequence and returns a Vector{ORF} containing the ORFs found in the sequence. It searches the sequence for start codons (ATG) and stops when it finds a stop codon (TAG, TAA, or TGA), adding each ORF it finds to the vector. The function also searches the reverse complement of the sequence, so it finds ORFs on both strands. This function has not ORFs size and overlapping condition contraints. Thus it might consider aa"M*" a posible encoding protein from the resulting ORFs. Extending the starting codons with the alternative_start = true will search for ATG, GTG, and TTG. Some studies have shown that in E. coli (K-12 strain), ATG, GTG and TTG are used 83 %, 14 % and 3 % respectively.

Keywords

alternative_start::Bool=false: If true will pass the extended start codons to search. This will increase 3x the exec. time.

min_len::Int64=6: Length of the allowed ORF. Default value allow aa"M*" a posible encoding protein from the resulting ORFs.

GeneFinder.orfgenerator — Method

orfgenerator(sequence::LongDNA; alternative_start::Bool=false, min_len = 6)

Generate ORFs from the given DNA sequence.

This function generates ORFs from the forward and reverse complement strands of the sequence using the locationgenerator function. It generates an ORF object for each range of indices returned by locationgenerator, and includes a '+' or '-' strand label to indicate the strand from which the ORF was generated. The sequence argument must be a LongDNA object, which is a type of DNA sequence with a longer maximum length than the DNA type.

Returns: A generator expression that yields ORF objects corresponding to the ORFs in the sequence.

GeneFinder.protein_generator — Method

protein_generator(sequence::LongDNA; kwargs...)
protein_generator(sequence::String; kwargs...)

As its name suggest this generator function iterates over the sequence to find proteins directly from a DNA sequence. The protein_generator function takes a LongDNA sequence and returns a Vector{CDS} containing the coding sequences (CDSs) found in the sequence and the associated ORF.

Keywords

code::GeneticCode=BioSequences.standard_genetic_code: The genetic code by which codons will be translated. See BioSequences.ncbi_trans_table for more info.

alternative_start::Bool=false: If true will pass the extended start codons to search. This will increase 3x the exec. time.

min_len::Int64=6: Length of the allowed ORF. Default value allow aa"M*" a posible encoding protein from the resulting ORFs.

GeneFinder.write_bed — Method

write_bed(file::String, seq::LongDNA; kwargs...)

Write BED data to a file.

Keywords

alternative_start::Bool=false: If true will pass the extended start codons to search. This will increase 3x the exec. time.

min_len::Int64=6: Length of the allowed ORF. Default value allow aa"M*" a posible encoding protein from the resulting ORFs.

GeneFinder.write_cds — Method

write_cds(file::String, seq::LongDNA; kwargs...)

Write a file containing the coding sequences (CDSs) of a given DNA sequence to the specified file.

Keywords

alternative_start: A boolean value indicating whether alternative start codons should be used when identifying CDSs. Default is false.

min_len: An integer representing the minimum length that a CDS must have in order to be included in the output file. Default is 6.

GeneFinder.write_proteins — Method

write_proteins(file::String, seq::LongDNA; kwargs...)

Write the protein sequences encoded by the coding sequences (CDSs) of a given DNA sequence to the specified file.

Keywords

code::GeneticCode=BioSequences.standard_genetic_code: The genetic code by which codons will be translated. See BioSequences.ncbi_trans_table for more info.

alternative_start::Bool=false: If true will pass the extended start codons to search. This will increase 3x the exec. time.

min_len::Int64=6: Length of the allowed ORF. Default value allow aa"M*" a posible encoding protein from the resulting ORFs.