The Main ORF type

The main type of the package is ORF which represents an Open Reading Frame.

GeneFinder.ORF — Type

struct ORF
    location::UnitRange{Int64}
    strand::Char
    frame::Int
    score::Union{Nothing, Float64}
end

The ORF struct represents an open reading frame in a DNA sequence. It has two fields:

location: which is a UnitRange{Int64} indicating the start and end locations of the ORF in the sequence

strand: is a Char type indicating whether the ORF is on the forward ('+') or reverse ('-') strand of the sequence.

frame: is an Int type indicating the reading frame of the ORF. The frame is the position of the first nucleotide of the codon that starts the ORF, relative to the start of the sequence. It can be 1, 2, or 3.

score: is a Union{Nothing, Float64} type indicating the score of the ORF. It can be a Float64 or nothing.

Finding ORFs

The function findorfs is the main function of the package. It is generic method that can handle different gene finding methods.

GeneFinder.findorfs — Function

findorfs(sequence::NucleicSeqOrView{DNAAlphabet{N}}; ::M, kwargs...) where {N, M<:GeneFinderMethod}

This is the main interface method for finding open reading frames (ORFs) in a DNA sequence.

It takes the following arguments:

sequence: The nucleic acid sequence to search for ORFs.

method: The algorithm used to find ORFs. It can be either NaiveFinder(), NaiveFinderScored() or yet other implementations.

Keyword Arguments regardless of the finder method:

alternative_start: A boolean indicating whether to consider alternative start codons. Default is false.

min_len: The minimum length of an ORF. Default is 6.

Keyword Arguments for NaiveFinderScored():

scoringscheme::BioMarkovChain: The scoring scheme to use for the scoring algorithm. Default is ECOLICDS.

Returns

A vector of ORF objects representing the found ORFs.

Example

sequence = randdnaseq(120)

    120nt DNA Sequence:
    GCCGGACAGCGAAGGCTAATAAATGCCCGTGCCAGTATC…TCTGAGTTACTGTACACCCGAAAGACGTTGTACGCATTT

findorfs(sequence, NaiveFinder())

    1-element Vector{ORF}:
    ORF(77:118, '-', 2, 0.0)

Finding ORFs using BioRegex and scoring

GeneFinder._locationiterator — Method

locationiterator(sequence::NucleicSeqOrView{DNAAlphabet{N}}; alternative_start::Bool=false) where {N}

This is an iterator function that uses regular expressions to search the entire ORF (instead of start and stop codons) in a LongSequence{DNAAlphabet{4}} sequence. It uses an anonymous function that will find the first regularly expressed ORF. Then using this anonymous function it creates an iterator that will apply it until there is no other CDS.

Note

As a note of the implementation we want to expand on how the ORFs are found:

The expression (?:[N]{3})*? serves as the boundary between the start and stop codons. Within this expression, the character class [N]{3} captures exactly three occurrences of any character (representing nucleotides using IUPAC codes). This portion functions as the regular codon matches. Since it is enclosed within a non-capturing group (?:) and followed by *?, it allows for the matching of intermediate codons, but with a preference for the smallest number of repetitions.

In summary, the regular expression ATG(?:[N]{3})*?T(AG|AA|GA) identifies patterns that start with "ATG," followed by any number of three-character codons (represented by "N" in the IUPAC code), and ends with a stop codon "TAG," "TAA," or "TGA." This pattern is commonly used to identify potential protein-coding regions within genetic sequences.

See more about the discussion here

GeneFinder.log_odds_ratio_decision_rule — Method

log_odds_ratio_decision_rule(
    sequence::LongSequence{DNAAlphabet{4}};
    codingmodel::BioMarkovChain,
    noncodingmodel::BioMarkovChain,
    η::Float64 = 1e-5
    )

Check if a given DNA sequence is likely to be coding based on a log-odds ratio. The log-odds ratio is a statistical measure used to assess the likelihood of a sequence being coding or non-coding. It compares the probability of the sequence generated by a coding model to the probability of the sequence generated by a non-coding model. If the log-odds ratio exceeds a given threshold (η), the sequence is considered likely to be coding. It is formally described as a decision rule:

\[S(X) = \log \left( \frac{{P_C(X_1=i_1, \ldots, X_T=i_T)}}{{P_N(X_1=i_1, \ldots, X_T=i_T)}} \right) \begin{cases} > \eta & \Rightarrow \text{{coding}} \\ < \eta & \Rightarrow \text{{noncoding}} \end{cases}\]

Arguments

sequence::NucleicSeqOrView{DNAAlphabet{N}}: The DNA sequence to be evaluated.

Keyword Arguments

codingmodel::BioMarkovChain: The transition model for coding regions, (default: ECOLICDS).

noncodingmodel::BioMarkovChain: The transition model for non-coding regions, (default: ECOLINOCDS)

η::Float64 = 1e-5: The threshold value (eta) for the log-odds ratio (default: 1e-5).

Returns

true if the sequence is likely to be coding.

false if the sequence is likely to be non-coding.

Raises

ErrorException: if the length of the sequence is not divisible by 3.

ErrorException: if the sequence contains a premature stop codon.

Example

sequence = dna"ATGGCATCTAG"
iscoding(sequence)  # Returns: true or false

GeneFinder.naivefinder — Method

naivefinder(sequence::NucleicAlphabet{DNAAlphabet{N}}; alternative_start::Bool=false, min_len::Int64=6)::Vector{ORF} where {N}

A simple implementation that finds ORFs in a DNA sequence.

The naivefinder function takes a LongSequence{DNAAlphabet{4}} sequence and returns a Vector{ORF} containing the ORFs found in the sequence. It searches entire regularly expressed CDS, adding each ORF it finds to the vector. The function also searches the reverse complement of the sequence, so it finds ORFs on both strands. Extending the starting codons with the alternative_start = true will search for ATG, GTG, and TTG. Some studies have shown that in E. coli (K-12 strain), ATG, GTG and TTG are used 83 %, 14 % and 3 % respectively.

Note

This function has not ORFs scoring scheme. Thus it might consider aa"M*" a posible encoding protein from the resulting ORFs.

Keywords

alternative_start::Bool=false: If true will pass the extended start codons to search. This will increase 3x the exec. time.

min_len::Int64=6: Length of the allowed ORF. Default value allow aa"M*" a posible encoding protein from the resulting ORFs.

GeneFinder.naivefinderscored — Method

naivefinderscored(sequence::NucleicSeqOrView{DNAAlphabet{N}}; alternative_start=false, min_len=6) where {N}

Find open reading frames (ORFs) in a nucleic acid sequence using a naive scoring algorithm. In this method the scoring is a log-odds ratio score defines as:

\[S(x) = \sum_{i=1}^{L} \beta_{x_{i}x} = \sum_{i=1} \log \frac{a^{\mathscr{m}_{1}}_{i-1} x_i}{a^{\mathscr{m}_{2}}_{i-1} x_i}\]

Arguments

sequence: The nucleic acid sequence to search for ORFs.

alternative_start: A boolean indicating whether to consider alternative start codons. Default is false.

min_len: The minimum length of an ORF to be considered. Default is 6.

scoringscheme: The scoring scheme to use for the scoring algorithm. Default is ECOLICDS.

Returns

A sorted vector of ORF objects representing the identified open reading frames.

Geting ORFs sequences

GeneFinder.getorfs — Method

getorfs(sequence::NucleicSeqOrView{DNAAlphabet{N}}, ::Alphabet, method::M; kwargs...) where {N, M<:GeneFinderMethod}

This function takes a NucleicSeqOrView{DNAAlphabet{N}} sequence and identifies the open reading frames (ORFs) using the findorfs() function under the selected finder method. The function then extracts the DNA or aminoacid sequence of each ORF and stores it in a Vector{LongSubSeq{A}}.

Arguments

sequence: The input sequence as a NucleicSeqOrView{DNAAlphabet{N}}

alphabet: The output sequence alphabet. It can be either DNAAlphabet{4}() or AminoAcidAlphabet().

method: The method used to find ORFs. It must be a subtype of GeneFinderMethod. (e.g., NaiveFinder())

Keyword Arguments

alternative_start::Bool=false: If set to true, the function considers alternative start codons when searching for ORFs. This increases the execution time by approximately 3x.

min_len::Int64=6: The minimum length of the allowed ORF. By default, it allows ORFs that can encode at least one amino acid (e.g., aa"M*").

Examples

sequence = randdnaseq(120)

    120nt DNA Sequence:
    GCCGGACAGCGAAGGCTAATAAATGCCCGTGCCAGTATC…TCTGAGTTACTGTACACCCGAAAGACGTTGTACGCATTT

getorfs(sequence, DNAAlphabet{4}(), NaiveFinder())

    1-element Vector{LongSubSeq{DNAAlphabet{4}}}:
     ATGCGTACAACGTCTTTCGGGTGTACAGTAACTCAGAGCTAA

Writing ORFs to files

GeneFinder.write_orfs_bed — Method

write_orfs_bed(input::NucleicSeqOrView{DNAAlphabet{N}}, output::Union{IOStream, IOBuffer}, method::M; kwargs...) where {N, M<:GeneFinderMethod}
write_orfs_bed(input::NucleicSeqOrView{DNAAlphabet{N}}, output::String, ::M; kwargs...) where {N, M<:GeneFinderMethod}
write_orfs_bed(input::String, output::Union{IOStream, IOBuffer}, method::M; kwargs...) where {M<:GeneFinderMethod}
write_orfs_bed(input::String, output::String, method::M; kwargs...) where {M<:GeneFinderMethod}

Write BED data to a file.

Arguments

input::NucleicAcidAlphabet{DNAAlphabet{N}}: The input DNA sequence.

output::IO: The otput format, it can be a file (String) or a buffer (IOStream or `IOBuffer)

method::M: The algorithm used to find ORFs. It can be either NaiveFinder() or NaiveFinderScored().

Keywords

alternative_start::Bool=false: If true, alternative start codons will be used when identifying CDSs. Default is false.

min_len::Int64=6: The minimum length that a CDS must have in order to be included in the output file. Default is 6.

GeneFinder.write_orfs_faa — Method

write_orfs_faa(input::NucleicSeqOrView{DNAAlphabet{4}}, output::Union{IOStream, IOBuffer}, method::M; kwargs...) where {N, M<:GeneFinderMethod}
write_orfs_faa(input::NucleicSeqOrView{DNAAlphabet{4}}, output::String, method::M; kwargs...) where {N, M<:GeneFinderMethod}
write_orfs_faa(input::String, output::Union{IOStream, IOBuffer}, method::M; kwargs...) where {M<:GeneFinderMethod}
write_orfs_faa(input::String, output::String, method::M; kwargs...) where {M<:GeneFinderMethod}

Write the protein sequences encoded by the coding sequences (CDSs) of a given DNA sequence to the specified file.

Arguments

input::NucleicAcidAlphabet{DNAAlphabet{N}}: The input DNA sequence.

output::IO: The otput format, it can be a file (String) or a buffer (IOStream or `IOBuffer)

method::M: The algorithm used to find ORFs. It can be either NaiveFinder() or NaiveFinderScored().

Keywords

code::GeneticCode=BioSequences.standard_genetic_code: The genetic code by which codons will be translated. See BioSequences.ncbi_trans_table for more info.

alternative_start::Bool=false: If true will pass the extended start codons to search. This will increase 3x the exec. time.

min_len::Int64=6: Length of the allowed ORF. Default value allow aa"M*" a posible encoding protein from the resulting ORFs.

Examples

filename = "output.faa"

seq = dna"ATGATGCATGCATGCATGCTAGTAACTAGCTAGCTAGCTAGTAA"

open(filename, "w") do file
     write_orfs_faa(seq, file)
end

GeneFinder.write_orfs_fna — Method

write_orfs_fna(input::NucleicSeqOrView{DNAAlphabet{N}}, output::Union{IOStream, IOBuffer}, method::M; kwargs...) where {N, M<:GeneFinderMethod}
write_orfs_fna(input::NucleicSeqOrView{DNAAlphabet{N}}, output::String, method::M; kwargs...) where {N, M<:GeneFinderMethod}
write_orfs_fna(input::String, output::Union{IOStream, IOBuffer}; kwargs...) where {M<:GeneFinderMethod}
write_orfs_fna(input::String, output::String, method::M; kwargs...) where {M<:GeneFinderMethod}

Write a file containing the coding sequences (CDSs) of a given DNA sequence to the specified file.

Arguments

input::NucleicAcidAlphabet{DNAAlphabet{N}}: The input DNA sequence.

output::IO: The otput format, it can be a file (String) or a buffer (IOStream or `IOBuffer)

method::M: The algorithm used to find ORFs. It can be either NaiveFinder() or NaiveFinderScored().

Keywords

alternative_start::Bool=false: If true, alternative start codons will be used when identifying CDSs. Default is false.

min_len::Int64=6: The minimum length that a CDS must have in order to be included in the output file. Default is 6.

Examples

filename = "output.fna"

seq = dna"ATGATGCATGCATGCATGCTAGTAACTAGCTAGCTAGCTAGTAA"

open(filename, "w") do file
     write_orfs_fna(seq, file)
end

GeneFinder.write_orfs_gff — Method

write_orfs_gff(input::NucleicSeqOrView{DNAAlphabet{N}}, output::Union{IOStream, IOBuffer}, method::M; kwargs...) where {N, M<:GeneFinderMethod}
write_orfs_gff(input::NucleicSeqOrView{DNAAlphabet{N}}, output::String, method::M; kwargs...) where {N, M<:GeneFinderMethod}
write_orfs_gff(input::String, output::Union{IOStream, IOBuffer}, method::M; kwargs...)  where {M<:GeneFinderMethod}
write_orfs_gff(input::String, output::String, method::M; kwargs...) where {M<:GeneFinderMethod}

Write GFF data to a file.

Arguments

input::NucleicAcidAlphabet{DNAAlphabet{N}}: The input DNA sequence.

output::IO: The otput format, it can be a file (String) or a buffer (IOStream or `IOBuffer)

method::M: The algorithm used to find ORFs. It can be either NaiveFinder() or NaiveFinderScored().

Keywords

code::GeneticCode=BioSequences.standard_genetic_code: The genetic code by which codons will be translated. See BioSequences.ncbi_trans_table for more info.

alternative_start::Bool=false: If true will pass the extended start codons to search. This will increase 3x the exec. time.

min_len::Int64=6: Length of the allowed ORF. Default value allow aa"M*" a posible encoding protein from the resulting ORFs.